feat(state): reversed order of manager stop
If managers are started in the order defined in the `se.managers` slice, I would expect them to be stopped in the reverse order. Why? Because maybe a manager M2 started after manager M1 depends on some functionality from M1, and stopping M1 before stopping M2 might lead to unexpected and undefined behavior.